pyproject.toml
[tool.poetry]
name = "planner"
version = "0.1.0"
description = ""
authors = ["Aleksey Eliseev <alekseyeliseev@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
black = "^24.8.0"
uvicorn = "^0.30.6"
pydantic = {extras = ["email"], version = "^2.8.2"}
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

